+++ oracle/atmospheres/marcs.py
@@ -1,11 +1,11 @@
-# coding: utf-8
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
 """ General utilities for parsing MARCS model atmospheres. """
 
 from __future__ import division, absolute_import, print_function
 
 __author__ = "Andy Casey <arc@ast.cam.ac.uk>"
-__all__ = ["parse_filename", "parse_photospheric_structure"]
 
 import gzip
 import numpy as np
