+++ oracle/atmospheres/pickler.py
@@ -1,4 +1,5 @@
-# coding: utf-8
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
 """ Convenience script to pickle a set of model atmospheres. """
